+++ dsmr_weather/tests/test_services.py
@@ -19,15 +19,13 @@ class TestDsmrWeatherServices(TestCase):
         self.schedule_process = ScheduledProcess.objects.get(module=settings.DSMRREADER_MODULE_WEATHER_UPDATE)
         self.schedule_process.update(active=True, planned=timezone.make_aware(timezone.datetime(2017, 1, 1)))
 
-    @mock.patch('dsmr_frontend.services.display_dashboard_message')
     @mock.patch('dsmr_weather.services.get_temperature_from_api')
     @mock.patch('django.utils.timezone.now')
-    def test_exception_handling(self, now_mock, get_temperature_from_api_mock, display_dashboard_message_mock):
+    def test_exception_handling(self, now_mock, get_temperature_from_api_mock):
         now_mock.return_value = timezone.make_aware(timezone.datetime(2017, 1, 1))
         get_temperature_from_api_mock.side_effect = AssertionError('TEST')  # Simulate any exception.
 
         dsmr_weather.services.run(self.schedule_process)
-        self.assertTrue(display_dashboard_message_mock.called)
 
         self.schedule_process.refresh_from_db()
         self.assertEqual(self.schedule_process.planned, timezone.now() + timezone.timedelta(hours=1))