+++ docs/changelog.rst
@@ -19,6 +19,7 @@ v4.1.0 - 2020-xx-xx
 
 - ``Changed``: Remote datalogger: Changed config to env vars [`#1057 <https://github.com/dennissiemensma/dsmr-reader/issues/1057>`_]
 - ``Changed``: Enabled retention by default for new installations [`#1000 <https://github.com/dennissiemensma/dsmr-reader/issues/1000>`_]
+- ``Changed``: Disabled display of Buienradar API errors on dashboard [`#1056 <https://github.com/dennissiemensma/dsmr-reader/issues/1056>`_]
 
 - ``Fixed``: Polyphase detection for Fluvius meters [`#1052 <https://github.com/dennissiemensma/dsmr-reader/issues/1052>`_]